Add JUnit tests Gradle and jar functionality
Added Testing with JUnit, Gradle, and jar functionality.
