+++ .gitignore
@@ -6,6 +6,7 @@
 # Gradle build files
 /.gradle/
 /build/
+/config/
 src/main/resources/docs/
 
 # MacOS custom attributes files created by Finder
