+++ build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'java'
     id 'application'
+    id 'checkstyle'
     id 'com.github.johnrengelman.shadow' version '7.1.2'
 }
 
@@ -28,11 +29,11 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("shaniqua.Shaniqua")
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveFileName = "Shaniqua.jar";
     archiveClassifier = null
 }
