--- src/main/java/Shaniqua/Command/CommandFailException.java
@@ -1,9 +1,0 @@
-package Shaniqua.Command;
-
-import Shaniqua.ShaniquaException;
-
-public class CommandFailException extends ShaniquaException {
-    CommandFailException(String msg) {
-        super("Command Failed: " + msg);
-    }
-}
