--- src/main/java/Shaniqua/Command/ModifyCommand.java
@@ -1,8 +1,0 @@
-package Shaniqua.Command;
-
-public abstract class ModifyCommand extends Command{
-    int idx;
-    public ModifyCommand(int idx) {
-        this.idx = idx;
-    }
-}
