--- src/main/java/Shaniqua/Command/StoreCommand.java
@@ -1,11 +1,0 @@
-package Shaniqua.Command;
-
-import Shaniqua.Storage.Storage;
-import Shaniqua.TaskCore.TaskList;
-import Shaniqua.UI.Ui;
-
-public class StoreCommand extends Command {
-    public void execute(TaskList tasks, Ui ui, Storage storage){
-
-    }
-}
