--- src/main/java/Shaniqua/Parser/ParserException.java
@@ -1,9 +1,0 @@
-package Shaniqua.Parser;
-
-import Shaniqua.ShaniquaException;
-
-public class ParserException extends ShaniquaException {
-    ParserException() {
-
-    }
-}
