--- src/main/java/Shaniqua/ShaniquaException.java
@@ -1,10 +1,0 @@
-package Shaniqua;
-
-public class ShaniquaException extends Exception {
-    public ShaniquaException() {
-        super();
-    }
-    public ShaniquaException(String msg) {
-        super(msg);
-    }
-}
