--- src/main/java/Shaniqua/Storage/StorageException.java
@@ -1,9 +1,0 @@
-package Shaniqua.Storage;
-
-import Shaniqua.ShaniquaException;
-
-public class StorageException extends ShaniquaException {
-    StorageException(String msg) {
-        super(msg);
-    }
-}
