--- src/main/java/Shaniqua/TaskCore/TaskListException.java
@@ -1,9 +1,0 @@
-package Shaniqua.TaskCore;
-
-import Shaniqua.ShaniquaException;
-
-public class TaskListException extends ShaniquaException {
-    TaskListException(String msg) {
-        super(msg);
-    }
-}
