--- src/main/java/Shaniqua/TaskCore/Tasks/FlexibleDateTimeException.java
@@ -1,9 +1,0 @@
-package Shaniqua.TaskCore.Tasks;
-
-import Shaniqua.ShaniquaException;
-
-public class FlexibleDateTimeException extends ShaniquaException {
-    FlexibleDateTimeException(String msg) {
-        super(msg);
-    }
-}
