+++ src/main/java/shaniqua/ShaniquaException.java
@@ -1,0 +1,10 @@
+package shaniqua;
+
+public class ShaniquaException extends Exception {
+    public ShaniquaException() {
+        super();
+    }
+    public ShaniquaException(String msg) {
+        super(msg);
+    }
+}
