+++ src/main/java/shaniqua/command/CommandFailException.java
@@ -1,0 +1,9 @@
+package shaniqua.command;
+
+import shaniqua.ShaniquaException;
+
+public class CommandFailException extends ShaniquaException {
+    CommandFailException(String msg) {
+        super("Command Failed: " + msg);
+    }
+}
