+++ src/main/java/shaniqua/command/ExitCommand.java
@@ -1,0 +1,11 @@
+package shaniqua.command;
+
+import shaniqua.storage.Storage;
+import shaniqua.taskcore.TaskList;
+import shaniqua.ui.Ui;
+
+public class ExitCommand extends Command {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        isExit = true;
+    }
+}
