+++ src/main/java/shaniqua/command/ModifyCommand.java
@@ -1,0 +1,8 @@
+package shaniqua.command;
+
+public abstract class ModifyCommand extends Command {
+    protected int idx;
+    public ModifyCommand(int idx) {
+        this.idx = idx;
+    }
+}
