+++ src/main/java/shaniqua/command/StoreCommand.java
@@ -1,0 +1,11 @@
+package shaniqua.command;
+
+import shaniqua.storage.Storage;
+import shaniqua.taskcore.TaskList;
+import shaniqua.ui.Ui;
+
+public class StoreCommand extends Command {
+    public void execute(TaskList tasks, Ui ui, Storage storage){
+
+    }
+}
