+++ src/main/java/shaniqua/parser/ParserException.java
@@ -1,0 +1,9 @@
+package shaniqua.parser;
+
+import shaniqua.ShaniquaException;
+
+public class ParserException extends ShaniquaException {
+    ParserException() {
+        super("Error parsing command");
+    }
+}
