+++ src/main/java/shaniqua/storage/StorageException.java
@@ -1,0 +1,9 @@
+package shaniqua.storage;
+
+import shaniqua.ShaniquaException;
+
+public class StorageException extends ShaniquaException {
+    StorageException(String msg) {
+        super(msg);
+    }
+}
