+++ src/main/java/shaniqua/taskcore/TaskListException.java
@@ -1,0 +1,9 @@
+package shaniqua.taskcore;
+
+import shaniqua.ShaniquaException;
+
+public class TaskListException extends ShaniquaException {
+    TaskListException(String msg) {
+        super(msg);
+    }
+}
