+++ src/main/java/shaniqua/taskcore/tasks/FlexibleDateTimeException.java
@@ -1,0 +1,9 @@
+package shaniqua.taskcore.tasks;
+
+import shaniqua.ShaniquaException;
+
+public class FlexibleDateTimeException extends ShaniquaException {
+    FlexibleDateTimeException(String msg) {
+        super(msg);
+    }
+}
