+++ src/main/java/shaniqua/taskcore/tasks/InvalidTaskDataException.java
@@ -1,0 +1,11 @@
+package shaniqua.taskcore.tasks;
+
+import shaniqua.ShaniquaException;
+
+public class InvalidTaskDataException extends ShaniquaException {
+    InvalidTaskDataException(String msg) {
+        super("Invalid task data entry: " + msg);
+    }
+    InvalidTaskDataException() {
+    }
+}
