+++ src/test/java/shaniqua/taskcore/TaskListTest.java
@@ -1,0 +1,4 @@
+package java.shaniqua.taskcore;
+
+public class TaskListTest {
+}
